requirements.txt
black==25.1.0
blinker==1.9.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
docstring-to-markdown==0.15
flake8==7.1.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
fpdf==1.7.2
greenlet==3.1.1
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
mysql==0.0.3
mysqlclient==2.2.7
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pylsp-mypy==0.7.0
pylsp-rope==0.1.17
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
pytoolconfig==1.3.1
pytz==2025.1
rope==1.13.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.39
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
Werkzeug==3.1.3
WTForms==3.2.1
